thread safe cooldown + different version check (not properly tested)
