+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>cz.gennario</groupId>
     <artifactId>RotatingHeads2</artifactId>
-    <version>1.2.0-FIXED-by-Len_137</version>
+    <version>1.2.1</version>
     <packaging>jar</packaging>
 
     <name>RotatingHeads2</name>
