+++ src/main/resources/plugin.yml
@@ -1,6 +1,6 @@
 name: RotatingHeads2
 author: Gennario
-version: 1.2.0-FIXED-by-Len_137
+version: 1.2.1
 main: cz.gennario.newrotatingheads.Main
 api-version: 1.13
 softdepend:
